@@ -1,13 +1,12 @@
+import { Chart } from '../chart/chart';
 import { Ranges } from '../chart/chart.component';
 import { DataMarksConfig } from './data-marks.config';
 
 export class DataMarks {
+  chart: Chart;
   config: DataMarksConfig;
   ranges: Ranges;
   setMethodsFromConfigAndDraw: () => void;
   resizeMarks: () => void;
   drawMarks: (transitionDuration: number) => void;
-  onPointerEnter: (event: PointerEvent) => void;
-  onPointerLeave: (event: PointerEvent) => void;
-  onPointerMove: (event: PointerEvent) => void;
 }
